+++ Commands/ClientSidePages/AddClientSidePage.cs
@@ -19,6 +19,10 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
         Code = @"PS:> Add-PnPClientSidePage -Name ""NewPage"" -ContentType ""MyPageContentType""",
         Remarks = "Creates a new Client-Side page named 'NewPage' and sets the content type to the content type specified",
         SortOrder = 2)]
+    [CmdletExample(
+        Code = @"PS:> Add-PnPClientSidePage -Name ""NewPageTemplate"" -PromoteAs Template",
+        Remarks = "Creates a new Client-Side page named 'NewPage' and saves as a template to the site.",
+        SortOrder = 2)]
     public class AddClientSidePage : PnPWebCmdlet
     {
         [Parameter(Mandatory = true, Position = 0, HelpMessage = "Specifies the name of the page.")]
@@ -68,9 +72,16 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
             // Create a page that persists immediately
             clientSidePage = SelectedWeb.AddClientSidePage(name);
             clientSidePage.LayoutType = LayoutType;
-            clientSidePage.Save(name);
+            if (PromoteAs == ClientSidePagePromoteType.Template)
+            {
+                clientSidePage.SaveAsTemplate(name);
+            }
+            else
+            {
+                clientSidePage.Save(name);
+            }
 
-            if(MyInvocation.BoundParameters.ContainsKey("ContentType"))
+            if (MyInvocation.BoundParameters.ContainsKey("ContentType"))
             {
                 ContentType ct = null;
                 if (ContentType.ContentType == null)