+++ Commands/ClientSidePages/GetClientSidePage.cs
@@ -12,11 +12,15 @@ namespace SharePointPnP.PowerShell.Commands.ClientSidePages
      [CmdletExample(
         Code = @"PS:> Get-PnPClientSidePage -Identity ""MyPage.aspx""",
         Remarks = "Gets the Modern Page (Client-Side) named 'MyPage.aspx' in the current SharePoint site",
-        SortOrder = 2)]
+        SortOrder = 1)]
     [CmdletExample(
         Code = @"PS:> Get-PnPClientSidePage ""MyPage""",
         Remarks = "Gets the Modern Page (Client-Side) named 'MyPage.aspx' in the current SharePoint site",
         SortOrder = 2)]
+    [CmdletExample(
+        Code = @"PS:> Get-PnPClientSidePage ""Templates/MyPageTemplate""",
+        Remarks = "Gets the Modern Page (Client-Side) named 'MyPageTemplate.aspx' from the templates folder of the Page Library in the current SharePoint site",
+        SortOrder = 3)]
     public class GetClientSidePage : PnPWebCmdlet
     {
         [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0, HelpMessage = "The name of the page")]