+++ Commands/Navigation/AddNavigationNode.cs
@@ -94,22 +94,35 @@ namespace SharePointPnP.PowerShell.Commands.Branding
                     if (Location == NavigationType.SearchNav)
                     {
                         nodeCollection = SelectedWeb.LoadSearchNavigation();
+#if !ONPREMISES
+                    }
+                    else if (Location == NavigationType.Footer)
+                    {
+                        nodeCollection = SelectedWeb.LoadFooterNavigation();
+#endif
                     }
                     else
                     {
                         nodeCollection = Location == NavigationType.QuickLaunch ? SelectedWeb.Navigation.QuickLaunch : SelectedWeb.Navigation.TopNavigationBar;
                         ClientContext.Load(nodeCollection);
                     }
-                    var addedNode = nodeCollection.Add(new NavigationNodeCreationInformation()
+                    if (nodeCollection != null)
                     {
-                        Title = Title,
-                        Url = Url,
-                        IsExternal = External.IsPresent,
-                        AsLastNode = !First.IsPresent
-                    });
-                    ClientContext.Load(addedNode);
-                    ClientContext.ExecuteQueryRetry();
-                    WriteObject(addedNode);
+                        var addedNode = nodeCollection.Add(new NavigationNodeCreationInformation()
+                        {
+                            Title = Title,
+                            Url = Url,
+                            IsExternal = External.IsPresent,
+                            AsLastNode = !First.IsPresent
+                        });
+                        ClientContext.Load(addedNode);
+                        ClientContext.ExecuteQueryRetry();
+                        WriteObject(addedNode);
+                    }
+                    else
+                    {
+                        throw new Exception("Navigation Node Collection is null");
+                    }
                 }
 #pragma warning restore CS0618 // Type or member is obsolete
             }