+++ Commands/Navigation/GetNavigationNode.cs
@@ -20,7 +20,7 @@ namespace SharePointPnP.PowerShell.Commands.Branding
         Remarks = @"Returns all navigation nodes in the quicklaunch navigation",
         SortOrder = 2)]
     [CmdletExample(
-        Code = @"PS:> Get-PnPNavigationNode -TopNavigationBar",
+        Code = @"PS:> Get-PnPNavigationNode -Location TopNavigationBar",
         Remarks = @"Returns all navigation nodes in the top navigation bar",
         SortOrder = 3)]
     [CmdletExample(
@@ -53,14 +53,23 @@ PS> $children = $node.Children",
                     if (Location == NavigationType.SearchNav)
                     {
                         navigationNodes = SelectedWeb.Navigation.GetNodeById(1040).Children;
+#if !ONPREMISES
+                    }
+                    else if (Location == NavigationType.Footer)
+                    {
+                        navigationNodes = SelectedWeb.LoadFooterNavigation();
+#endif
                     }
                     else
                     {
                         navigationNodes = Location == NavigationType.QuickLaunch ? SelectedWeb.Navigation.QuickLaunch : SelectedWeb.Navigation.TopNavigationBar;
                     }
-                    var nodesCollection = ClientContext.LoadQuery(navigationNodes);
-                    ClientContext.ExecuteQueryRetry();
-                    WriteObject(GetTree(nodesCollection, 0));
+                    if (navigationNodes != null)
+                    {
+                        var nodesCollection = ClientContext.LoadQuery(navigationNodes);
+                        ClientContext.ExecuteQueryRetry();
+                        WriteObject(GetTree(nodesCollection, 0));
+                    }
                 }
                 else
                 {
@@ -82,10 +91,20 @@ PS> $children = $node.Children",
                                 nodes = SelectedWeb.Navigation.GetNodeById(1040).Children;
                                 break;
                             }
+#if !ONPREMISES
+                        case NavigationType.Footer:
+                            {
+                                nodes = SelectedWeb.LoadFooterNavigation();
+                                break;
+                            }
+#endif
                     }
-                    ClientContext.Load(nodes);
-                    ClientContext.ExecuteQueryRetry();
-                    WriteObject(nodes, true);
+                    if (nodes != null)
+                    {
+                        ClientContext.Load(nodes);
+                        ClientContext.ExecuteQueryRetry();
+                        WriteObject(nodes, true);
+                    }
                 }
             }
             if (MyInvocation.BoundParameters.ContainsKey("Id"))