+++ Commands/Provisioning/Site/ReadProvisioningTemplate.cs
@@ -7,13 +7,13 @@ using SharePointPnP.PowerShell.Commands.Utilities;
 using System;
 using System.IO;
 using System.Management.Automation;
+using System.Text;
 
 namespace SharePointPnP.PowerShell.Commands.Provisioning
 {
-
     [Cmdlet(VerbsCommunications.Read, "PnPProvisioningTemplate")]
     [Alias("Load-PnPProvisioningTemplate")]
-    [CmdletHelp("Loads/Reads a PnP file from the file system",
+    [CmdletHelp("Loads/Reads a PnP file from the file system or a string",
         Category = CmdletHelpCategory.Provisioning)]
     [CmdletExample(
        Code = @"PS:> Read-PnPProvisioningTemplate -Path template.pnp",
@@ -23,25 +23,60 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
        Code = @"PS:> Read-PnPProvisioningTemplate -Path template.pnp -TemplateProviderExtensions $extensions",
        Remarks = "Loads a PnP file from the file system using some custom template provider extensions while loading the file.",
        SortOrder = 2)]
+    [CmdletExample(
+       Code = @"PS:> Read-PnPProvisioningTemplate -Xml $xml",
+       Remarks = "Reads a PnP Provisioning template from a string containing the XML of a provisioning template",
+       SortOrder = 3)]
     public class ReadProvisioningTemplate : PSCmdlet
     {
-        [Parameter(Mandatory = true, Position = 0, HelpMessage = "Filename to read from, optionally including full path.")]
+        const string ParameterSet_PATH = "By Path";
+        const string ParameterSet_XML = "By XML";
+
+        [Parameter(Mandatory = true, Position = 0, ParameterSetName = ParameterSet_PATH, HelpMessage = "Filename to read from, optionally including full path.")]
         public string Path;
 
-        [Parameter(Mandatory = false, HelpMessage = "Allows you to specify ITemplateProviderExtension to execute while loading the template.")]
+        [Parameter(Mandatory = true, Position = 1, ParameterSetName = ParameterSet_XML, HelpMessage = "Variable to read from, containing the valid XML of a provisioning template.")]
+        public string Xml;
+
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets, HelpMessage = "Allows you to specify ITemplateProviderExtension to execute while loading the template.")]
         public ITemplateProviderExtension[] TemplateProviderExtensions;
 
         protected override void ProcessRecord()
         {
-            if(MyInvocation.InvocationName.ToLower() == "load-pnpprovisioningtemplate")
+            if (MyInvocation.InvocationName.ToLower() == "load-pnpprovisioningtemplate")
             {
                 WriteWarning("Load-PnPProvisioningTemplate has been deprecated in favor of Read-PnPProvisioningTemplate which supports the same parameters.");
             }
-            if (!System.IO.Path.IsPathRooted(Path))
+            switch (ParameterSetName)
             {
-                Path = System.IO.Path.Combine(SessionState.Path.CurrentFileSystemLocation.Path, Path);
+                case ParameterSet_PATH:
+                    {
+
+                        if (!System.IO.Path.IsPathRooted(Path))
+                        {
+                            Path = System.IO.Path.Combine(SessionState.Path.CurrentFileSystemLocation.Path, Path);
+                        }
+                        WriteObject(LoadProvisioningTemplateFromFile(Path, TemplateProviderExtensions));
+                        break;
+                    }
+                case ParameterSet_XML:
+                    {
+                        WriteObject(LoadProvisioningTemplateFromString(Xml, TemplateProviderExtensions));
+                        break;
+                    }
             }
-            WriteObject(LoadProvisioningTemplateFromFile(Path, TemplateProviderExtensions));
+        }
+
+        internal static ProvisioningTemplate LoadProvisioningTemplateFromString(string xml, ITemplateProviderExtension[] templateProviderExtensions)
+        {
+            var formatter = new XMLPnPSchemaFormatter();
+
+            XMLTemplateProvider provider = new XMLStreamTemplateProvider();
+
+            // Get the XML document from a File Stream
+            Stream stream = new MemoryStream(Encoding.UTF8.GetBytes(xml));
+
+            return provider.GetTemplate(stream);
         }
 
         internal static ProvisioningTemplate LoadProvisioningTemplateFromFile(string templatePath, ITemplateProviderExtension[] templateProviderExtensions)
@@ -75,7 +110,6 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning
             {
                 provider = new XMLFileSystemTemplateProvider(fileConnector.Parameters[FileConnectorBase.CONNECTIONSTRING] + "", "");
             }
-
             ProvisioningTemplate provisioningTemplate = provider.GetTemplate(templateFileName, templateProviderExtensions);
             provisioningTemplate.Connector = provider.Connector;